v1
In this work we compare the accuracy of three different classification algorithms in two benchmark image datasets, the MNIST and CIFAR10 datasets. We benchmarked the use of K-Nearest Neighbor, Multi-Layer Perceptron and Convolutional Neural Networks. We discuss how inter and intra-class differences affect classification accuracies as well as discuss how each of the algorithms deals with these differences.